+++ packages-content-model/roosterjs-content-model-core/lib/coreApi/getDOMSelection.ts
@@ -1,4 +1,3 @@
-import { isSelectionReverted } from 'roosterjs-content-model-dom';
 import type {
     DOMSelection,
     GetDOMSelection,
@@ -31,4 +30,17 @@ function getNewSelection(core: StandaloneEditorCore): DOMSelection | null {
               isReverted: isSelectionReverted(selection),
           }
         : null;
+}
+
+function isSelectionReverted(selection: Selection | null | undefined): boolean {
+    if (selection && selection.rangeCount > 0) {
+        const range = selection.getRangeAt(0);
+        return (
+            !range.collapsed &&
+            selection.focusNode != range.endContainer &&
+            selection.focusOffset != range.endOffset
+        );
+    }
+
+    return false;
 }
