+++ packages-content-model/roosterjs-content-model-dom/lib/modelToDom/contentModelToDom.ts
@@ -1,6 +1,5 @@
 import toArray from '../domUtils/toArray';
 import { isNodeOfType } from '../domUtils/isNodeOfType';
-import { isSelectionReverted } from '../domUtils/isSelectionReverted';
 import type {
     ContentModelDocument,
     DOMSelection,
@@ -112,4 +111,17 @@ function calcPosition(
     }
 
     return result;
+}
+
+function isSelectionReverted(selection: Selection | null | undefined): boolean {
+    if (selection && selection.rangeCount > 0) {
+        const range = selection.getRangeAt(0);
+        return (
+            !range.collapsed &&
+            selection.focusNode != range.endContainer &&
+            selection.focusOffset != range.endOffset
+        );
+    }
+
+    return false;
 }
